backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
pandas==2.2.2
numpy==1.26.4
requests==2.31.0
python-dotenv==1.0.1
gunicorn==22.0.0
gevent==24.2.1